Updated Vector3 to support "minecraft trigonometry", and implemented Location class to provide entity location methods (See #4) Added command "dbgpos" for debugging client position and yaw Still plenty of work to be done in this area yet.

+++ Chraft/Persistence/ClientSurrogate.cs
@@ -13,8 +13,8 @@ namespace Chraft.Persistence
 		public double X { get; set; }
 		public double Y { get; set; }
 		public double Z { get; set; }
-		public float Yaw { get; set; }
-		public float Pitch { get; set; }
+		public double Yaw { get; set; }
+		public double Pitch { get; set; }
 		public Inventory Inventory { get; set; }
 	}
 }
